+++ bmlib/src/bmlib_memory.h
@@ -31,7 +31,7 @@ typedef struct bm_memcpy_info {
 	u64 src_device_addr;
 	u64 device_addr;
 	union {
-		u32 size;
+		u64 size;
 		struct {
 			u16 width;
 			u16 height;
@@ -82,7 +82,7 @@ bm_status_t bm_memcpy_s2d_poll(bm_handle_t handle, bm_device_mem_t dst, void *sr
 bm_status_t sg_memcpy_s2d_poll(bm_handle_t handle, sg_device_mem_t dst, void *src);
 bm_status_t bm_memcpy_s2d_poll_u64(bm_handle_t handle, bm_device_mem_u64_t dst, void *src);
 void *bm_mem_get_system_addr(struct bm_mem_desc mem);
-u32 bm_mem_get_size(struct bm_mem_desc mem);
+u64 bm_mem_get_size(struct bm_mem_desc mem);
 u64 sg_mem_get_size(struct sg_mem_desc mem);
 u64 bm_mem_get_size_u64(struct bm_mem_desc_u64 mem);
 bm_status_t bm_mem_mmap_device_mem(bm_handle_t handle, bm_device_mem_t *dmem, u64 *vmem);
